PR Revision: Rename the new Apple Health data provider and add a new requiresV2Api property on data type definitions.
