+++ src/helpers/daily-data-providers/apple-health-steps-while-wearing-device.ts
